% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AA57OLMatrix}
\alias{AA57OLMatrix}
\title{Overlapping Matrix for 57 amino acid and secondary structure combinations.}
\format{A matrix with 57 by 57 elements.}
\usage{
AA57OLMatrix
}
\description{
A dataset containing the pre-calculated overlapping matrix 57 amino acid and secondary structure combinations:
}
\details{
\itemize{
  \item column names or row names.
    \itemize{
      \item First letter. amino acid typings in single-letter format, not include glycine ("A", "C", "D", "E", "F", "H", "I", "K", "L", "M", "N", "P", "Q", "R", "S", "T", "V", "W", "Y")
      \item second letter. secondary structure in single-letter format ("B", "H", "C")
    }
}
}
\keyword{datasets}
