% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AA57OLWeights}
\alias{AA57OLWeights}
\title{Overlapping weights for 57 amino acid and secondary structure combinations.}
\format{A vector of 57 elements.}
\usage{
AA57OLWeights
}
\description{
A vector containing the pre-calculated overlapping weights 57 amino acid and secondary structure combinations:
}
\details{
\itemize{
  \item names.
    \itemize{
      \item First letter. amino acid typings in single-letter format, not include glycine ("A", "C", "D", "E", "F", "H", "I", "K", "L", "M", "N", "P", "Q", "R", "S", "T", "V", "W", "Y")
      \item second letter. secondary structure in single-letter format ("B", "H", "C")
    }
}
}
\keyword{datasets}
