% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aaCodes1Letter19}
\alias{aaCodes1Letter19}
\title{Single-letter amino acid naming convention.}
\format{A vector of 19 elements.}
\usage{
aaCodes1Letter19
}
\description{
A vector containing amino acid names with all letters capitalized, (no glycine), they are:
"A", "C", "D", "E", "F", "H", "I", "K", "L", "M", "N", "P", "Q", "R", "S", "T", "V", "W", "Y".
}
\keyword{datasets}
