% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aaCodes3Letter1stCap}
\alias{aaCodes3Letter1stCap}
\title{Three-letter amino acid naming convention with first letter capitalized.}
\format{A vector of 19 elements.}
\usage{
aaCodes3Letter1stCap
}
\description{
A vector containing amino acid names with first letter capitalized (no glycine), they are:
"Ala", "Cys", "Asp", "Glu", "Phe", "His", "Ile", "Lys", "Leu", "Met", "Asn", "Pro", "Gln", "Arg", "Ser", "Thr", "Val", "Trp", "Tyr".
}
\keyword{datasets}
