% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aaCodes3LetterAllCap.v}
\alias{aaCodes3LetterAllCap.v}
\alias{aaCodes3LetterAllCap}
\title{Three-letter amino acid naming convention with all letters capitalized.}
\format{A vector of 19 elements.}
\usage{
aaCodes3LetterAllCap.v
}
\description{
A vector containing amino acid names with all letters capitalized (no glycine), they are:
"ALA", "CYS", "ASP", "GLU", "PHE", "HIS", "ILE", "LYS", "LEU", "MET", "ASN", "PRO", "GLN", "ARG", "SER", "THR", "VAL", "TRP", "TYR".
}
\keyword{datasets}
