% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aaFreq}
\alias{aaFreq}
\title{Pre-defined amino acid frequency data.}
\format{A data frame with 57 rows and 2 columns.}
\usage{
aaFreq
}
\description{
Sample data for testing \code{calculate_MSE()} function.
\itemize{
  \item AA_SS. amino acid typing and secondary structure naming combination (single-letter convention)
  \item Freq. relative cumulated frequency
}
}
\keyword{datasets}
