% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AvgCov.t}
\alias{AvgCov.t}
\title{Average covariance across three secondary structures for all amino acid typings (including oxydized cystine).}
\format{A data frame with 20 rows and 2 columns.}
\usage{
AvgCov.t
}
\description{
A dataset containing average covariance across three secondary structure for all amino acid typings (including oxydized cystine):
\itemize{
  \item AA. amino acid typing names
  \item AvgCov. average covariance value across three secondary structures
}
}
\keyword{datasets}
