% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CarbonCov.t}
\alias{CarbonCov.t}
\title{Covariance values of chemical shifts of alpha and beta carbons.}
\format{A data frame with 79 rows and 2 columns.}
\usage{
CarbonCov.t
}
\description{
A dataset containing covariance values of chemical shifts of alpha and beta carbons for each amino acid across three secondary structure for all amino acid typings (including oxidized cystine but not glycine):
\itemize{
  \item AA.SS. amino acid typing and secondary structure combination (single-letter convention)
  \item Value. covariance value
}
}
\keyword{datasets}
