% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CASdTable}
\alias{CASdTable}
\title{Standard deviation of chemical shift values for alpha carbon .}
\format{A data frame with 21 rows and 5 columns.}
\usage{
CASdTable
}
\description{
A dataset containing Standard deviation of chemical shift values for alpha carbon for each amino acid across three secondary structure for all amino acid typings (including oxydized cystine and glycine):
\itemize{
  \item Residue. amino acid typing name (single-letter convention)
  \item C. coil
  \item H. helix
  \item B. beta strand
  \item A. average of three secondary structures.
}
}
\keyword{datasets}
