% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CBSdTable}
\alias{CBSdTable}
\title{Standard deviation of chemical shift values for beta carbon .}
\format{A data frame with 21 rows and 5 columns.}
\usage{
CBSdTable
}
\description{
A dataset containing Standard deviation of chemical shift values for beta carbon for each amino acid across three secondary structure for all amino acid typings (including oxidized cystine and glycine):
\itemize{
  \item Residue. amino acid typing (single-letter convention)
  \item C. coil
  \item H. helix
  \item B. beta strand
  \item A. average of three secondary structures.
}
}
\keyword{datasets}
