% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chemicalShifts}
\alias{chemicalShifts}
\title{Pre-defined sample chemical shifts data.}
\format{A data frame with 55 rows and 4 columns.}
\usage{
chemicalShifts
}
\description{
Sample data for testing \code{calculate_MSE()} function.
\itemize{
  \item AA. amino acid typing (single-letter convention)
  \item SS. secondary structure (single-letter convention)
  \item CA. chemical shift value of alpha carbon
  \item CB. chemical shift value of beta carbon
}
}
\keyword{datasets}
