% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ID}
\alias{ID}
\title{RefDB ID included in the BaMORC package.}
\format{A vector of 1557 elements.}
\usage{
ID
}
\description{
A vector containing all the RefDB ID included in the BaMORC package.
}
\keyword{datasets}
