% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{class}
\name{inversedMatrices}
\alias{inversedMatrices}
\title{inversedMatrices.}
\format{An object of inversed matrices with their accessing functions.}
\usage{
inversedMatrices
}
\description{
A data set containing all the pre-calculated inversed matrices and the access functions.
\describe{
 \item{$getInvMatrix(name)}{get the inversed matrix basing on the name, which is the amino acid and secondary structure combination (single-letter naming convention)}
 \item{$getNames()}{get all the names of the inverse matrices}
 \item{$Matrices}{the data set of all the pre-calculated inversed matricies}
}
}
\keyword{datasets}
