% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{jpred_fetcher}
\alias{jpred_fetcher}
\title{Using JPred Mass-submission scheduler program to submit protein sequence and return secodary structure results.}
\usage{
jpred_fetcher(protein_sequence)
}
\arguments{
\item{protein_sequence}{protein sequence}
}
\value{
protein secondary structure information
}
\description{
Using JPred Mass-submission scheduler program to submit protein sequence and return secodary structure results.
}
\examples{
library(jpredapir)
protein_sequence <- "MQVWPIEGIKKFETLSYLPPLTVEDLLKQI"
\dontrun{secondary_structure <- jpred_fetcher(protein_sequence)}
}
