% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{read_db_file}
\alias{read_db_file}
\title{\code{read_db_file()} reads in data from existing database that included in the BaMORC package. This database was extracted from RefDB database.}
\usage{
read_db_file(id)
}
\arguments{
\item{id}{BMRB or RefDB entry ID.}
}
\value{
Protein sequence, secondary structure information and chemical shifts dataframe.
}
\description{
\code{read_db_file()} reads in data from existing database that included in the BaMORC package. This database was extracted from RefDB database.
}
\examples{
id = 4022
head(read_db_file(id))
}
