% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RefDB.StatCA}
\alias{RefDB.StatCA}
\title{Statistics of chemical shifts values of alpha carbons from RefDB.}
\format{A data frame with 21 rows and 9 columns.}
\usage{
RefDB.StatCA
}
\description{
A dataset containing statistics of chemical shifts of alpha carbons for each amino acid from RefDB data:
\itemize{
  \item Residue. amino acid typing (single-letter convention)
  \item Coil.mu. mean value of alpha carbon in coil
  \item Coil.sd. mean value of alpha carbon in coil
  \item Helix.mu. mean value of alpha carbon in helix
  \item Helix.sd. standard deviation value of alpha carbon in helix
  \item Beta.mu. mean value of alpha carbon in coil
  \item Beta.sd. standard deviation value of alpha carbon in coil
  \item Avg.mu. average of mean values of alpha carbon across three secondary structures
  \item Avg.sd. average of standard deviation values of alpha carbon across three secondary structures
}
}
\keyword{datasets}
