% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RefDB.StatCB}
\alias{RefDB.StatCB}
\title{Statistics of chemical shifts values of beta carbons from RefDB.}
\format{A data frame with 21 rows and 9 columns.}
\usage{
RefDB.StatCB
}
\description{
A dataset containing statistics of chemical shifts of beta carbons for each amino acid from RefDB data:
\itemize{
  \item Residue. amino acid typing (single-letter convention)
  \item Coil.mu. mean value of beta carbon in coil
  \item Coil.sd. mean value of beta carbon in coil
  \item Helix.mu. mean value of beta carbon in helix
  \item Helix.sd. standard deviation value of beta carbon in helix
  \item Beta.mu. mean value of beta carbon in coil
  \item Beta.sd. standard deviation value of beta carbon in coil
  \item Avg.mu. average of mean values of beta carbon across three secondary structures
  \item Avg.sd. average of standard deviation values of beta carbon across three secondary structures
}
}
\keyword{datasets}
