% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{class}
\name{RefDB_data}
\alias{RefDB_data}
\title{RefDB object}
\format{A class contains 1557 13C datasets and access functions.}
\usage{
RefDB_data
}
\description{
Collection of data from Re-referenced Protein Chemical shift Database (RefDB) and their access functions.
}
\keyword{datasets}
