% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barmorc.R
\name{calculate_chi_squared_stat}
\alias{calculate_chi_squared_stat}
\title{Calculates a chi squared statistic(s).}
\usage{
calculate_chi_squared_stat(cacb_pair)
}
\arguments{
\item{cacb_pair}{A pair of carbon chemical shifts \code{c(Ca, Cb)}}
}
\value{
A list of chi-squared statistics basing the combination of amino acid typings and secondary structures.
}
\description{
\code{calculate_chi_squared_stat} Given a pair of C_alpha and C_beta chemical shifts, this function will return a list of calculated chisquare statistics based on the combination of amino acid typings and secondary structures.
}
\examples{
calculate_chi_squared_stat(c(54,45))
}
