% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{read_DB_File}
\alias{read_DB_File}
\title{\code{read_DB_file()} reads in data from existing database that included in the BaMORC package. This database was extracted from RefDB database.}
\usage{
read_DB_File(id)
}
\arguments{
\item{id}{BMRB or RefDB entry ID.}
}
\value{
Protein sequence, secondary structure information and chemical shifts dataframe.
}
\description{
\code{read_DB_file()} reads in data from existing database that included in the BaMORC package. This database was extracted from RefDB database.
}
\examples{
id = 4022
head(read_DB_File(id))
}
