% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{read_NMRSTAR_file}
\alias{read_NMRSTAR_file}
\title{Extracts data from BMRB STAR 3.0 file.
\code{read_NMRSTAR_file()} parses BMRB STAR 3.0 file. It will extract sequence information and chemical shifts for both alpha and beta carbons.}
\usage{
read_NMRSTAR_file(file_path)
}
\arguments{
\item{file_path}{File path where input chemical shifts file is located}
}
\value{
Protein sequence and chemical shifts dataframe.
}
\description{
Extracts data from BMRB STAR 3.0 file.
\code{read_NMRSTAR_file()} parses BMRB STAR 3.0 file. It will extract sequence information and chemical shifts for both alpha and beta carbons.
}
\examples{
## Download a BMRB file
library(BMRBr)
bmrb_download(4020, output_dir = "./")

## Read in BMRB file and procec
file_path = "bmr4020.str"
head(read_NMRSTAR_file(file_path))

## Delete downloaded BMRB file
unlink("./bmr4020.str")

}
