% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_file_generator.R
\name{sample_data_generator}
\alias{sample_data_generator}
\title{sample_data_generator}
\usage{
sample_data_generator(input_type, output_dir = "")
}
\arguments{
\item{input_type}{flag indicates what sample data will be generated}
}
\value{
save a sample file in a temporary location for demonstration
}
\description{
Generate sample data file for testing purpose. Depending the input_type, the following sample files will be generated:
\itemize{
  \item{ssc_sample}{bpti_HNcoCACB.txt, for testing \code{unassigned_bamorc()} function}
  \item{csv}{sample_input.csv, sample file in csv format}
  \item{sc}{sample_input_sc.txt, sample file with semicolon as delimiter}
  \item{ws}{sample_input_ws.txt, sample file with white space as delimiter}
}
}
\examples{
input_type = "ws" 
output_dir = tempdir()
sample_data_generator(input_type = input_type, output_dir=output_dir)
head(read_raw_file(paste0(output_dir, "/sample_input_ws.txt"), delim="ws"))
unlink(paste0(output_dir, "/sample_input_ws.txt"))

input_type = "csv"
output_dir = tempdir()
sample_data_generator(input_type = input_type, output_dir=output_dir)
head(read_raw_file(paste0(output_dir, "/sample_input.csv"), delim="comma"))
unlink(paste0(output_dir, "/sample_input.csv"))

input_type = "sc"
output_dir = tempdir()
sample_data_generator(input_type = input_type, output_dir=output_dir)
head(read_raw_file(paste0(output_dir, "/sample_input_sc.txt"), delim="semicolon"))
unlink(paste0(output_dir, "/sample_input_sc.txt"))
}
